Add logback config file
So that debug messages are not printed out when executing tests.

Logback logging config file is "redundant" since we already have
log4j.properties. The former is used by test process and is preferred
since it log4j can not be found on DSE classpath. The latter is used by
obsolete EmbeddedCassandra and needs to be removed in the future.